Changing text color on hover proposal
Since the background goes to a kind of light yellow, if the color of the text is in white and it will not be readable, this can be little bit confusing. I would like to suggest to take into consideration using a grey color.

I have chosen that grey in order to have the same value as the one used in the last update date element at the end of the blog. But  a lighter grey can match too.